refactor(diff-viewer): split App and EditableDiffPane into cohesive hooks
Both components carried `no-giant-component` suppressions with a note saying the
debt was accepted because neither could be split without making the code worse.
They can. `App` went from 1006 lines to 247 and `EditableDiffPane` from 486 to
214, so both suppressions are deleted rather than relocated onto whatever ended
up nearest, along with `prefer-useReducer` and the note itself.

The seams are cohesion, not line count:

  diffViewerHost.ts        the conversation with the extension host
  diffViewerModel.ts       the pure derivation: segments, layout, highlighting
  diffOverlayPainter.ts    everything that measures or draws against the DOM
  diffViewerScroll.ts      both scroll axes and the per-document reset
  diffViewerInput.ts       what the keyboard can do: refusals, change-to-change nav
  editableDraftSession.ts  the draft an edit is composed in

`useRevertHunk` is separate from `useDiffViewerHost` because reverting needs the
rendered segments, which are derived from the payload that hook delivers, so
folding it in would force the model to run before the hook producing its input.

Two components come out of the render. `DiffPaneColumn` was the same 39 lines
written twice, differing only in `left`/`right` down to the `key` prefix and the
test id -- the shape where one side quietly stops matching the other.
`DiffViewerToolbar` follows so `App` clears the ceiling with room to take an edit
rather than landing five lines under it.

A ref that crosses a hook boundary stops looking stable to
`react-hooks/exhaustive-deps`: a `useRef` declared in the same component is
treated as stable, while the same ref returned from a custom hook is an ordinary
value. The refs that moved are now listed in the dependency arrays that read
them, which is a runtime no-op. The painter's unmount cleanup moved into the hook
that owns the frame handle, since reading it anywhere else warns that the value
will have changed by the time the cleanup runs.
